nave guardada en base de datos

+++ src/logica/Usuario.java
@@ -7,7 +7,7 @@ public class Usuario {
 	public String nombreUsuario;//nombre que es usuario decida ponerse
 	public String contrasenya;//contraseña del usuario
 	
-	public double nivel;//nivel del jugador
+	public int nivel;//nivel del jugador
 	public ArrayList<Partida> part;//Lista de partidas del jugador
 
 	/**Crea un usuario
@@ -72,14 +72,14 @@ public class Usuario {
 	/**
 	 * @return nivel Nivel del usuario
 	 */
-	public double getNivel() {
+	public int getNivel() {
 		return nivel;
 	}
 
 	/**Modifica el nivel del usuario
 	 * @param nivel Nuevo nivel del usuario
 	 */
-	public void setNivel(double nivel) {
+	public void setNivel(int nivel) {
 		this.nivel = nivel;
 	}
 }
